feat(application): make all commands and queries return Result

+++ Birds.Application/Commands/DeleteBird/DeleteBirdCommandHandler.cs
@@ -1,14 +1,18 @@
-﻿using Birds.Application.Interfaces;
+﻿using Birds.Application.Common.Models;
+using Birds.Application.Interfaces;
 using Birds.Application.Notifications;
 using MediatR;
 
 namespace Birds.Application.Commands.DeleteBird
 {
     public class DeleteBirdCommandHandler(IBirdRepository repository, IUnitOfWork unitOfWork, IMediator mediator)
-        : IRequestHandler<DeleteBirdCommand>
+        : IRequestHandler<DeleteBirdCommand, Result>
     {
-        public async Task Handle(DeleteBirdCommand request, CancellationToken cancellationToken)
+        public async Task<Result> Handle(DeleteBirdCommand request, CancellationToken cancellationToken)
         {
+            if (request is null)
+                return Result.Failure("Request canno be null");
+
             var bird = await repository.GetByIdAsync(request.Id, cancellationToken);
 
             repository.Remove(bird);
@@ -16,6 +20,8 @@ namespace Birds.Application.Commands.DeleteBird
             await unitOfWork.SaveChangesAsync(cancellationToken);
 
             await mediator.Publish(new BirdDeletedNotification(bird.Id), cancellationToken);
+
+            return Result.Success();
         }
     }
 }
